refactor: 将 Tauri 功能命令归回现有资源模块
